+++ src/commands/status.ts
@@ -74,7 +74,11 @@ export async function run(args: ParsedArgs): Promise<void> {
     throw err;
   });
 
-  const config = resolveConfig(root, undefined, { verbose: !!verbose });
+  const config = resolveConfig(
+    root,
+    { main: args.main, dev: args.dev, remote: args.remote },
+    { verbose: !!verbose }
+  );
   const current = await getCurrentBranch(root, {
     dryRun: !!dryRun,
     verbose: !!verbose,